requirements.txt
fastapi
uvicorn
python-dotenv
google-cloud-firestore
langchain
openai
langchain-openai
langchain-community
faiss-cpu
chromadb
sentence-transformers
numpy
